+++ xbmc/FileItem.h
@@ -28,7 +28,7 @@
 #include "guilib/GUIListItem.h"
 #include "utils/Archive.h"
 #include "utils/ISerializable.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 #include "SortFileItem.h"
 #include "utils/LabelFormatter.h"
 #include "GUIPassword.h"
