+++ xbmc/TextureDatabase.cpp
@@ -22,7 +22,7 @@
 #include "TextureDatabase.h"
 #include "utils/log.h"
 #include "utils/Crc32.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 
 CTextureDatabase::CTextureDatabase()
 {
