+++ xbmc/XBDateTime.cpp
@@ -19,7 +19,7 @@
  *
  */
 
-#include "DateTime.h"
+#include "XBDateTime.h"
 #include "LangInfo.h"
 #include "guilib/LocalizeStrings.h"
