+++ xbmc/filesystem/MythDirectory.h
@@ -22,7 +22,7 @@
 
 #include "IDirectory.h"
 #include "MythSession.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 
 namespace XFILE
 {
