+++ xbmc/filesystem/MythFile.cpp
@@ -20,7 +20,7 @@
  */
 
 #include "MythFile.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 #include "FileItem.h"
 #include "utils/URIUtils.h"
 #include "DllLibCMyth.h"
