+++ xbmc/utils/TimeUtils.cpp
@@ -20,7 +20,7 @@
  */
 
 #include "TimeUtils.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 
 #ifdef __APPLE__
 #ifdef __ppc__
