+++ xbmc/video/windows/GUIWindowFullScreen.cpp
@@ -48,7 +48,7 @@
 #include "threads/SingleLock.h"
 #include "utils/log.h"
 #include "utils/TimeUtils.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 #include "input/ButtonTranslator.h"
 
 #include <stdio.h>
