+++ xbmc/win32/XBMC_PC.cpp
@@ -24,7 +24,7 @@
 #include "WIN32Util.h"
 #include "shellapi.h"
 #include "dbghelp.h"
-#include "DateTime.h"
+#include "XBDateTime.h"
 #include "threads/Thread.h"
 #include "Application.h"
